[function] Move `CFunction2` over
- Move `CFunction2` over to its own dedicated source file
