+++ src/main/scala/com/nec/spark/agile/filter/FilterFunction.scala
@@ -1,9 +1,8 @@
 package com.nec.spark.agile.filter
 
-import com.nec.spark.agile.core.FunctionTemplateTrait
+import com.nec.spark.agile.core.{CFunction2, FunctionTemplateTrait}
+import com.nec.spark.agile.core.CFunction2.CFunctionArgument.{Pointer, PointerPointer}
 import com.nec.spark.agile.CExpressionEvaluation.CodeLines
-import com.nec.spark.agile.CFunction2
-import com.nec.spark.agile.CFunction2.CFunctionArgument.{Pointer, PointerPointer}
 import com.nec.spark.agile.CFunctionGeneration.{CExpression, CVector, VeFilter}
 
 object FilterFunction {
