+++ src/main/scala/com/nec/spark/agile/projection/ProjectionFunction.scala
@@ -1,11 +1,11 @@
 package com.nec.spark.agile.projection
 
 import com.nec.spark.agile.CExpressionEvaluation.CodeLines
-import com.nec.spark.agile.CFunction2
-import com.nec.spark.agile.CFunction2.CFunctionArgument
 import com.nec.spark.agile.CFunctionGeneration._
 import com.nec.spark.agile.StringProducer.FrovedisStringProducer
-import com.nec.spark.agile.core.FunctionTemplateTrait
+import com.nec.spark.agile.core.{CFunction2, FunctionTemplateTrait}
+import com.nec.spark.agile.core.CFunction2.CFunctionArgument
+
 
 case class ProjectionFunction(name: String,
                               data: List[CVector],
