+++ src/test/scala/com/nec/ve/eval/RealExpressionEvaluationUtils.scala
@@ -21,8 +21,8 @@ package com.nec.ve.eval
 
 import com.nec.arrow.WithTestAllocator
 import com.nec.cmake.eval.OldUnifiedGroupByFunctionGeneration
-import com.nec.spark.agile.CFunction2
 import com.nec.spark.agile.CFunctionGeneration._
+import com.nec.spark.agile.core.CFunction2
 import com.nec.spark.agile.filter.FilterFunction
 import com.nec.spark.agile.projection.ProjectionFunction
 import com.nec.spark.agile.sort.SortFunction
